requirements.txt
argparse
colorama
munch
pyyaml